refactor: move converting method to StatisticsCacheUtils
rename StatisticsUtils to StatisticsCacheUtils
move converting method to StatisticsCacheUtils
- convert sumsMap and countsMap to GenderEmotionAmountAverageDto
- convert sumsMap to GenderDailyAmountSumDto
- convert sumsMap and countsMap to GenderSatisfactionAverageDto
move casting method to StatisticsCacheUtils
